requirements.txt
streamlit
pytesseract
Pillow
openai
whisper
torch
pandas
notion-client
dropbox
